fix: Break handlerbarHelper into submodules
