+++ src/handlebar/index.ts
@@ -1,7 +1,13 @@
-import { HandlerBarHelper } from "./handleBarDefinition";
 import registerCustomHandlebars from "./loadCustomHandlebars";
 import logger from "../logger";
 import { CsvHelper } from "./CsvHelper";
+import { NowHelper } from "./NowHelper";
+import { RandomValueHelper } from "./RandomValueHelper";
+import { RequestHelper } from "./RequestHelper";
+import { NumBetweenHelper } from "./NumBetweenHelper";
+import { FileHelper } from "./FileHelper";
+import { CodeHelper } from "./CodeHelper";
+import { InjectHelper } from "./InjectHelper";
 /**
  * Creates a instance of HandleBarHelper and register each custom helper
  * - If external helper is null, do not call registerCustomHandlebars()
@@ -9,16 +15,22 @@ import { CsvHelper } from "./CsvHelper";
  */
 const registerHandlebars = (extHelpers: string) => {
   logger.info("Handlebar helpers registration started");
-  const handlerBarHelper = new HandlerBarHelper();
   const csvHelper = new CsvHelper();
-  handlerBarHelper.nowHelper();
-  handlerBarHelper.randomValueHelper();
-  handlerBarHelper.requestHelper();
-  handlerBarHelper.numBetweenHelper();
-  handlerBarHelper.fileHelper();
-  handlerBarHelper.codeHelper();
-  handlerBarHelper.injectHelper();
-  csvHelper.csvHelper();
+  const nowHelper = new NowHelper();
+  const randomValueHelper = new RandomValueHelper();
+  const requestHelper = new RequestHelper();
+  const numBetweenHelper = new NumBetweenHelper();
+  const fileHelper = new FileHelper();
+  const codeHelper = new CodeHelper();
+  const injectHelper = new InjectHelper();
+  nowHelper.register();
+  randomValueHelper.register();
+  requestHelper.register();
+  numBetweenHelper.register();
+  fileHelper.register();
+  codeHelper.register();
+  injectHelper.register();
+  csvHelper.register();
   if (extHelpers !== null) {
     registerCustomHandlebars(extHelpers);
   }
